(refactor) moved actions to shared folder
