+++ src/applications/check-in/actions/navigation/index.js
