+++ src/applications/check-in/actions/navigation/navigation.actions.unit.spec.js
